--- mcp_tools/network_recon/core_network_scanning.py
@@ -1,158 +1,0 @@
-# mcp_tools/core_network_scanning.py
-
-
-from typing import Dict, Any
-
-
-def register_core_network_scanning_tools(mcp, hexstrike_client, logger, HexStrikeColors):
-    @mcp.tool()
-    def whois_lookup(target: str) -> Dict[str, Any]:
-        """
-        Perform a WHOIS lookup for a domain or IP address.
-
-        Args:
-            target: The domain or IP to query
-
-        Returns:
-            WHOIS lookup results
-        """
-        try:
-            response = hexstrike_client.safe_post("api/tools/whois", {"target": target})
-            return response
-        except Exception as e:
-            logger.error(f"WHOIS lookup failed: {e}")
-            return {"error": str(e)}
-    
-    @mcp.tool()
-    def nmap_scan(target: str, scan_type: str = "-sV", ports: str = "", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute an enhanced Nmap scan against a target with real-time logging.
-
-        Args:
-            target: The IP address or hostname to scan
-            scan_type: Scan type (e.g., -sV for version detection, -sC for scripts)
-            ports: Comma-separated list of ports or port ranges
-            additional_args: Additional Nmap arguments
-
-        Returns:
-            Scan results with enhanced telemetry
-        """
-        data: Dict[str, Any] = {
-            "target": target,
-            "scan_type": scan_type,
-            "ports": ports,
-            "additional_args": additional_args
-        }
-        logger.info(f"{HexStrikeColors.FIRE_RED}🔍 Initiating Nmap scan: {target}{HexStrikeColors.RESET}")
-
-        # Use enhanced error handling by default
-        data["use_recovery"] = True
-        result = hexstrike_client.safe_post("api/tools/nmap", data)
-
-        if result.get("success"):
-            logger.info(f"{HexStrikeColors.SUCCESS}✅ Nmap scan completed successfully for {target}{HexStrikeColors.RESET}")
-
-            # Check for recovery information
-            if result.get("recovery_info", {}).get("recovery_applied"):
-                recovery_info = result["recovery_info"]
-                attempts = recovery_info.get("attempts_made", 1)
-                logger.info(f"{HexStrikeColors.HIGHLIGHT_YELLOW} Recovery applied: {attempts} attempts made {HexStrikeColors.RESET}")
-        else:
-            logger.error(f"{HexStrikeColors.ERROR}❌ Nmap scan failed for {target}{HexStrikeColors.RESET}")
-
-            # Check for human escalation
-            if result.get("human_escalation"):
-                logger.error(f"{HexStrikeColors.CRITICAL} HUMAN ESCALATION REQUIRED {HexStrikeColors.RESET}")
-
-        return result
-
-    @mcp.tool()
-    def gobuster_scan(url: str, mode: str = "dir", wordlist: str = "/usr/share/wordlists/dirb/common.txt", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute Gobuster to find directories, DNS subdomains, or virtual hosts with enhanced logging.
-
-        Args:
-            url: The target URL
-            mode: Scan mode (dir, dns, fuzz, vhost)
-            wordlist: Path to wordlist file
-            additional_args: Additional Gobuster arguments
-
-        Returns:
-            Scan results with enhanced telemetry
-        """
-        data: Dict[str, Any] = {
-            "url": url,
-            "mode": mode,
-            "wordlist": wordlist,
-            "additional_args": additional_args
-        }
-        logger.info(f"{HexStrikeColors.CRIMSON}📁 Starting Gobuster {mode} scan: {url}{HexStrikeColors.RESET}")
-
-        # Use enhanced error handling by default
-        data["use_recovery"] = True
-        result = hexstrike_client.safe_post("api/tools/gobuster", data)
-
-        if result.get("success"):
-            logger.info(f"{HexStrikeColors.SUCCESS}✅ Gobuster scan completed for {url}{HexStrikeColors.RESET}")
-
-            # Check for recovery information
-            if result.get("recovery_info", {}).get("recovery_applied"):
-                recovery_info = result["recovery_info"]
-                attempts = recovery_info.get("attempts_made", 1)
-                logger.info(f"{HexStrikeColors.HIGHLIGHT_YELLOW} Recovery applied: {attempts} attempts made {HexStrikeColors.RESET}")
-        else:
-            logger.error(f"{HexStrikeColors.ERROR}❌ Gobuster scan failed for {url}{HexStrikeColors.RESET}")
-
-            # Check for alternative tool suggestion
-            if result.get("alternative_tool_suggested"):
-                alt_tool = result["alternative_tool_suggested"]
-                logger.info(f"{HexStrikeColors.HIGHLIGHT_BLUE} Alternative tool suggested: {alt_tool} {HexStrikeColors.RESET}")
-
-        return result
-
-    @mcp.tool()
-    def nuclei_scan(target: str, severity: str = "", tags: str = "", template: str = "", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute Nuclei vulnerability scanner with enhanced logging and real-time progress.
-
-        Args:
-            target: The target URL or IP
-            severity: Filter by severity (critical,high,medium,low,info)
-            tags: Filter by tags (e.g. cve,rce,lfi)
-            template: Custom template path
-            additional_args: Additional Nuclei arguments
-
-        Returns:
-            Scan results with discovered vulnerabilities and telemetry
-        """
-        data: Dict[str, Any] = {
-            "target": target,
-            "severity": severity,
-            "tags": tags,
-            "template": template,
-            "additional_args": additional_args
-        }
-        logger.info(f"{HexStrikeColors.BLOOD_RED}🔬 Starting Nuclei vulnerability scan: {target}{HexStrikeColors.RESET}")
-
-        # Use enhanced error handling by default
-        data["use_recovery"] = True
-        result = hexstrike_client.safe_post("api/tools/nuclei", data)
-
-        if result.get("success"):
-            logger.info(f"{HexStrikeColors.SUCCESS}✅ Nuclei scan completed for {target}{HexStrikeColors.RESET}")
-
-            # Enhanced vulnerability reporting
-            if result.get("stdout") and "CRITICAL" in result["stdout"]:
-                logger.warning(f"{HexStrikeColors.CRITICAL} CRITICAL vulnerabilities detected! {HexStrikeColors.RESET}")
-            elif result.get("stdout") and "HIGH" in result["stdout"]:
-                logger.warning(f"{HexStrikeColors.FIRE_RED} HIGH severity vulnerabilities found! {HexStrikeColors.RESET}")
-
-            # Check for recovery information
-            if result.get("recovery_info", {}).get("recovery_applied"):
-                recovery_info = result["recovery_info"]
-                attempts = recovery_info.get("attempts_made", 1)
-                logger.info(f"{HexStrikeColors.HIGHLIGHT_YELLOW} Recovery applied: {attempts} attempts made {HexStrikeColors.RESET}")
-        else:
-            logger.error(f"{HexStrikeColors.ERROR}❌ Nuclei scan failed for {target}{HexStrikeColors.RESET}")
-
-        return result
